[DF-2720] Removed inheritance from BaseObject at Component

+++ Sources/Internal/Scene3D/Components/LodComponent.cpp
@@ -320,7 +320,7 @@ void LodComponent::SetLayerVisibility(int32 layerNum, bool visible)
 {
 	DVASSERT(0 <= layerNum && layerNum < MAX_LOD_LAYERS);
 
-	if(lodLayers.size() > layerNum)
+	if((int32)lodLayers.size() > layerNum)
 	{
 		int32 size = lodLayers[layerNum].nodes.size();
 		for (int32 i = 0; i < size; i++)
